+++ src/main/java/persistence/entity/persistencecontext/Snapshot.java
@@ -3,7 +3,7 @@ package persistence.entity.persistencecontext;
 import java.util.HashMap;
 import java.util.Map;
 
-import static persistence.sql.dml.clause.PrimaryKeyValue.getPrimaryKeyValue;
+import static persistence.sql.ddl.clause.primkarykey.PrimaryKeyValue.getPrimaryKeyValue;
 
 public class Snapshot {
     private final Map<EntityKey, Object> snapshot;
