+++ src/main/java/persistence/entity/persister/EntityPersister.java
@@ -11,7 +11,7 @@ import persistence.sql.dml.querybuilder.UpdateQueryBuilder;
 import java.lang.reflect.Field;
 import java.util.Arrays;
 
-import static persistence.sql.dml.clause.PrimaryKeyValue.getPrimaryKeyValue;
+import static persistence.sql.ddl.clause.primkarykey.PrimaryKeyValue.getPrimaryKeyValue;
 
 public class EntityPersister {
     private final JdbcTemplate jdbcTemplate;
